backend/requirements.txt
# FastAPI & Web
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.31
alembic==1.13.2
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis & Caching
redis==5.0.7

# Celery
celery==5.4.0

# Validation & Serialization
pydantic==2.8.2
pydantic-settings==2.3.4
email-validator==2.2.0

# HTTP Client
httpx==0.27.0

# AWS
boto3==1.34.0

# ML & AI
tensorflow==2.16.1
mediapipe==0.10.14
numpy==1.26.4
scikit-learn==1.5.1

# Monitoring
prometheus-client==0.20.0
sentry-sdk[fastapi]==2.11.0

# Testing
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0
factory-boy==3.3.0
faker==26.0.0

# Development
black==24.4.2
isort==5.13.2
flake8==7.1.0
mypy==1.11.0
pre-commit==3.7.1

# Utilities
python-dateutil==2.9.0
pytz==2024.1
Pillow==10.4.0
python-magic==0.4.27
cryptography==42.0.8
